Added test suite to avoid creation of orchestrator in each test class.
